refactor(scheduler): rename m_mock_scheduler → mock_scheduler
