+++ packages/devextreme/js/__internal/scheduler/__tests__/agenda.test.ts
@@ -3,7 +3,7 @@ import {
 } from '@jest/globals';
 
 import { createScheduler } from './__mock__/create_scheduler';
-import { setupSchedulerTestEnvironment } from './__mock__/m_mock_scheduler';
+import { setupSchedulerTestEnvironment } from './__mock__/mock_scheduler';
 
 describe('agenda', () => {
   it('should render appointment', async () => {
