+++ packages/devextreme/js/__internal/scheduler/__tests__/toolbar_adaptivity.test.ts
@@ -3,7 +3,7 @@ import {
 } from '@jest/globals';
 
 import { createScheduler } from './__mock__/create_scheduler';
-import { setupSchedulerTestEnvironment } from './__mock__/m_mock_scheduler';
+import { setupSchedulerTestEnvironment } from './__mock__/mock_scheduler';
 
 describe('Toolbar Adaptivity', () => {
   it('should have viewSwitcher with locateInMenu: "auto" by default', async () => {
