+++ packages/devextreme/js/__internal/scheduler/appointment_popup/appointment_popup.integration.test.ts
@@ -9,7 +9,7 @@ import { toMilliseconds } from '@ts/utils/toMilliseconds';
 
 import fx from '../../../common/core/animation/fx';
 import { createScheduler } from '../__tests__/__mock__/create_scheduler';
-import { setupSchedulerTestEnvironment } from '../__tests__/__mock__/m_mock_scheduler';
+import { setupSchedulerTestEnvironment } from '../__tests__/__mock__/mock_scheduler';
 import { DEFAULT_SCHEDULER_OPTIONS } from '../utils/options/constants';
 
 const CLASSES = {
